moving tests to own folder
