+++ package.json
@@ -10,7 +10,7 @@
     "build-grpc": "grpc_tools_node_protoc --plugin=protoc-gen-ts_proto=./node_modules/.bin/protoc-gen-ts_proto --ts_proto_out=./src/grpc/out/ --ts_proto_opt=outputServices=nice-grpc,outputServices=generic-definitions,useExactTypes=false,esModuleInterop=true --proto_path=./src/grpc/proto/ src/grpc/proto/wallet-manager-grpc-service.proto",
     "build": "npm run build-grpc && tsc",
     "watch": "tsc -w",
-        "lint": "eslint --ignore-path .gitignore --ext .js,.ts .",
+    "lint": "eslint --ignore-path .gitignore --ext .js,.ts .",
     "prettier": "prettier --ignore-path .gitignore --write .",
     "test": "jest --silent=false"
   },
