+++ src/wallets/index.ts
@@ -82,3 +82,4 @@ export function createWalletToolbox(
       throw new Error(`Unknown chain name ${chainName}`);
   }
 }
+
