+++ tsconfig.json
@@ -21,10 +21,11 @@
   },
   "include": [
     "src"
-  ],
+, "test/balances/sui.test.ts"],
   "exclude": [
     "node_modules",
     "lib/**/*",
-    "src/**/*.test.ts"
+    "src/**/*.test.ts",
+    "test"
   ]
 }
